re-use low-rank for RGCN
trigger ci
